src/category/domain/entities/category.ts
//TDD - Kent Beck
// Tests: Fail -> Success -> Refactor

export type CategoryProps = {
    name: string;
    description?: string;
    is_active?: boolean;
    created_at?: Date;
}

export class Category {
    constructor(public readonly props: CategoryProps) {}

    get name() {
        return this.props.name
    }

    get description() {
        return this.props.description
    }

    get is_active() {
        return this.props.is_active
    }

    get created_at() {
        return this.props.created_at
    }
}

// const category = new Category({name: 'Movie'})

